refactor: add model and repository for computer
